requirements.txt
fastapi==0.63.0
pydantic==1.8.1
numpy==1.18.1
requests==2.23.0
joblib==1.0.1
loguru==0.4.1
xgboost==1.1.1
uvicorn==0.11.7
python-dotenv==0.17.0
google-cloud-storage==1.37.1
scikit-learn==0.24.1